Replacing existing work history flow
