--- app/views/_includes/review/work-history-2.njk
@@ -1,151 +1,0 @@
-{% set workHistory = applicationValue(["work-history"]) or {} %}
-{% set workHistoryDecision = applicationValue(["work-history-decision"]) %}
-
-{% if workHistoryDecision == "full-time-education" %}
-
-  {{ govukSummaryList({
-    classes: "app-summary",
-    rows: [{
-      key: {
-        text: "Do you have any work history?"
-      },
-      value: {
-        text: "No, I have always been in full time education"
-      },
-      actions: {
-        items: [{
-          href: applicationPath + "/work-history-2",
-          text: "Change",
-          visuallyHiddenText: "explanation"
-        }]
-      }
-    }]
-  }) }}
-
-{% elseif workHistoryDecision == "no" %}
-
-  {{ govukSummaryList({
-    classes: "app-summary",
-    rows: [{
-      key: {
-        text: "Explanation of why you’ve been out of the workplace"
-      },
-      value: {
-        text: (applicationValue(["work-history", "missing"]) or "Not entered") | nl2br | safe
-      },
-      actions: {
-        items: [{
-          href: applicationPath + "/work-history-2",
-          text: "Change",
-          visuallyHiddenText: "explanation"
-        }]
-      }
-    }]
-  }) }}
-
-{% else %}
-
-  {% set history = workHistory | toArray | selectattr("id") | sort(attribute="start-date") %}
-
-    <section class="app-summary-card govuk-!-margin-bottom-6">
-    <div class="app-summary-card__body">
-    <table class="govuk-table">
-      <caption class="govuk-table__caption govuk-heading-m">Jobs and breaks in work history</caption>
-      <thead class="govuk-table__head">
-        <tr class="govuk-table__row">
-          <th class="govuk-table__header">From</th>
-          <th class="govuk-table__header">To</th>
-          <th class="govuk-table__header">Description</th>
-
-          <th class="govuk-table__header app-table__cell--action">Actions</th>
-        </tr>
-      </thead>
-      <tbody>
-
-        {% if history.length >= 1 %}
-          {% for item in history %}
-
-          {# periodStart = Earliest date in required work history period #}
-          {% set periodStart = "now" | date | replace("2019", "2014") %}
-          {% set periodStartEpoch = periodStart | date("x") | int %}
-
-          {# firstStart = Start date of first item in work history #}
-          {% set firstStart = item["start-date"] %}
-          {% set firstStartEpoch = firstStart | date("x") | int %}
-
-            <tr>
-              <td class="govuk-table__cell"><span class="app-date">{{ item["start-date"] | date('LLL yyyy') }}</span></td>
-              <td class="govuk-table__cell">
-              {% if item["end-date-present"] == "yes" %}
-                Present
-              {% else %}
-                <span class="app-date">{{ item["end-date"] | date("LLL yyyy") if item["end-date"] != "now" else "Present" }}</span>
-              {% endif %}
-              </td>
-              <td class="govuk-table__cell">
-                {% if item["category"] == "break" %}
-                  <b>Break in work history</b>
-                  <br><br>
-                  {{ item["description"] }}
-                {% else %}
-
-
-
-                  <p class="govuk-body">
-                    <b class="govuk-!-font-weight-bold">{{ item["role"] }}</b>
-                    {% if item["type"] == "Part time" %}(Part time){% endif %}
-                  <br>{{ item["org"] }}
-                </p>
-
-                  {% if item["worked-with-children"] == "Yes" %}
-                    <p class="govuk-body-s govuk-!-margin-bottom-1">This role involved working with children</p>
-                  {% endif %}
-                {% endif %}
-              </td>
-              <td class="govuk-table__cell app-table__cell--action">
-
-
-                {% set editPath = (applicationPath + "/work-history-2/break/" + item.id) if item["category"] == "break" else (applicationPath + "/work-history-2/" + item.id) %}
-
-
-                <a class="govuk-link" href="{{ editPath }}">Change</a><br>
-                <a class="govuk-link app-link--destructive" href="{{ applicationPath + "/work-history-2/" + item.id + "/remove" }}">Remove</a>
-              </td>
-            </tr>
-
-            {# Unexplained break (during work history) #}
-            {% set nextItem = history[loop.index] %}
-            {% set itemStart = item["end-date"] %}
-            {% set itemStartEpoch = itemStart | date("x") | int %}
-            {% set itemEnd = nextItem["start-date"] if nextItem else "now" %}
-            {% set itemEndEpoch = itemEnd | date("x") | int %}
-            {% set itemDuration = itemEndEpoch - itemStartEpoch %}
-            {% if itemEndEpoch > itemStartEpoch and itemDuration > 2678400000 %}
-
-             <tr class="govuk-table__row" id="work-history-break-{{ itemStart | date("yyyy-LL") }}">
-              <td class="govuk-table__cell app-table__cell--break-prompt" colspan="4">
-                You have a break in your work history ({{ itemDuration | duration }})
-
-                <br><a href="{{ applicationPath + "/work-history-2/" + newId + "?start=" + itemStart | date("yyyy-LL") + "&end=" + itemEnd | date("yyyy-LL") }} ">Add another job</a> or
-
-                <a href="{{ applicationPath + "/work-history-2/break/" + newId + "?start=" + itemStart | date("yyyy-LL") + "&end=" + itemEnd | date("yyyy-LL") }} ">add a reason for this break</a>
-
-              </tr>
-            </td>
-
-            {% endif %}
-
-
-          {% endfor %}
-        {% else %}
-          <tr class="govuk-table__row">
-            <td colspan="5" class="govuk-table__cell">
-              No work history added yet.
-            </td>
-          </tr>
-        {% endif %}
-      </tbody>
-    </table>
-  </div>
-  </section>
-{% endif %}
