--- app/views/_includes/review/work-history-3.njk
@@ -1,142 +1,0 @@
-{% set workHistory = applicationValue(["work-history"]) or {} %}
-{% set workHistoryDecision = applicationValue(["work-history-decision"]) %}
-
-{% if workHistoryDecision == "full-time-education" %}
-
-  {{ govukSummaryList({
-    classes: "app-summary",
-    rows: [{
-      key: {
-        text: "Do you have any work history?"
-      },
-      value: {
-        text: "No, I have always been in full time education"
-      },
-      actions: {
-        items: [{
-          href: applicationPath + "/work-history-2",
-          text: "Change",
-          visuallyHiddenText: "explanation"
-        }]
-      }
-    }]
-  }) }}
-
-{% elseif workHistoryDecision == "no" %}
-
-  {{ govukSummaryList({
-    classes: "app-summary",
-    rows: [{
-      key: {
-        text: "Explanation of why you’ve been out of the workplace"
-      },
-      value: {
-        text: (applicationValue(["work-history", "missing"]) or "Not entered") | nl2br | safe
-      },
-      actions: {
-        items: [{
-          href: applicationPath + "/work-history-2",
-          text: "Change",
-          visuallyHiddenText: "explanation"
-        }]
-      }
-    }]
-  }) }}
-
-{% else %}
-
-  {% set history = workHistory | toArray | selectattr("id") | sort(attribute="start-date") %}
-
-    <section class="govuk-!-margin-bottom-6">
-
-        {% if history.length >= 1 %}
-          <hr class="govuk-section-break govuk-section-break--visible govuk-section-break--m ">
-          {% for item in history %}
-
-          {# periodStart = Earliest date in required work history period #}
-          {% set periodStart = "now" | date | replace("2019", "2014") %}
-          {% set periodStartEpoch = periodStart | date("x") | int %}
-
-          {# firstStart = Start date of first item in work history #}
-          {% set firstStart = item["start-date"] %}
-          {% set firstStartEpoch = firstStart | date("x") | int %}
-
-          <div class="govuk-grid-row">
-          <div class="govuk-grid-column-two-thirds">
-
-          <h2 class="govuk-heading-s govuk-!-margin-bottom-0">
-            {% if item['category'] == 'break' %}
-              Break in work history
-            {% else %}
-              {{  item['org'] }}
-            {%  endif %}
-          </h2>
-          <p class="govuk-body govuk-!-margin-bottom-0">
-            {% if item['category'] == 'break' %}
-              {{ item["description"] }}
-            {% else %}
-              {{ item["role"] }}
-              {% if item["type"] == "Part time" %}(Part time){% endif %}
-            {% endif %}
-          </p>
-          <p class="govuk-body">
-            {{ item["start-date"] | date('LLL yyyy') }}
-
-            {% if item['start-date-approx'] | length  > 0 %}
-              (estimate)
-            {% endif %}
-
-            –
-            {% if item["end-date-present"] == "yes" %}
-              Present
-            {% else %}
-              {{ item["end-date"] | date("LLL yyyy") if item["end-date"] != "now" else "Present" }}
-
-            {% if item['end-date-approx'] | length  > 0 %}
-              (estimate)
-            {% endif %}
-
-            {% endif %}
-          </p>
-          {% if item["worked-with-children"] == "Yes" %}
-            <p class="govuk-body-s govuk-!-margin-bottom-1">This role used skills relevant to teaching</p>
-          {% endif %}
-
-          {% set editPath = (applicationPath + "/work-history-2/break/" + item.id) if item["category"] == "break" else (applicationPath + "/work-history-2/" + item.id) %}
-
-
-          </div>
-          <div class="govuk-grid-column-one-third govuk-body app-grid-column-one-third--actions">
-            <a class="govuk-link" href="{{ editPath }}">Change</a>&nbsp;&nbsp;&nbsp;
-            <a class="govuk-link" href="{{ applicationPath + "/work-history-2/" + item.id + "/remove" }}">Delete</a>
-          </div>
-          </div>
-
-          <hr class="govuk-section-break govuk-section-break--visible govuk-section-break--m ">
-            {# Unexplained break (during work history) #}
-            {% set nextItem = history[loop.index] %}
-            {% set itemStart = item["end-date"] %}
-            {% set itemStartEpoch = itemStart | date("x") | int %}
-            {% set itemEnd = nextItem["start-date"] if nextItem else "now" %}
-            {% set itemEndEpoch = itemEnd | date("x") | int %}
-            {% set itemDuration = itemEndEpoch - itemStartEpoch %}
-            {% if itemEndEpoch > itemStartEpoch and itemDuration > 2678400000 %}
-
-              <div class="govuk-inset-text app-inset-text--important govuk-!-margin-top-0 govuk-!-margin-bottom-0">
-                You have a break in your work history ({{ itemDuration | duration }})
-
-                <br><a href="{{ applicationPath + "/work-history-2/" + newId + "?start=" + itemStart | date("yyyy-LL") + "&end=" + itemEnd | date("yyyy-LL") }} ">Add another job</a> or
-
-                <a href="{{ applicationPath + "/work-history-2/break/" + newId + "?start=" + itemStart | date("yyyy-LL") + "&end=" + itemEnd | date("yyyy-LL") }} ">add a reason for this break</a>
-
-              </div>
-                <hr class="govuk-section-break govuk-section-break--visible govuk-section-break--m ">
-
-            {% endif %}
-
-
-          {% endfor %}
-        {% endif %}
-
-  </section>
-{% endif %}
