+++ app/views/application/index.njk
@@ -204,9 +204,9 @@
 
 
     {% if applicationValue("work-history-decision") %}
-      {% set workHistoryPath = applicationPath + "/work-history-2/review" %}
+      {% set workHistoryPath = applicationPath + "/work-history/review" %}
     {% else %}
-      {% set workHistoryPath = applicationPath + "/work-history-2" %}
+      {% set workHistoryPath = applicationPath + "/work-history" %}
     {% endif %}
