+++ app/views/application/review.njk
@@ -91,7 +91,7 @@
 
   <section class="govuk-!-margin-bottom-8">
     <h2 class="govuk-heading-m govuk-!-font-size-27">Work experience</h2>
-    {% include "_includes/review/work-history-3.njk" %}
+    {% include "_includes/review/work-history.njk" %}
     {% include "_includes/review/school-experience.njk" %}
   </section>
