--- app/views/application/work-history-2/index.njk
@@ -1,66 +1,0 @@
-{% extends "_layout.njk" %}
-{% set applicationPath = "/application/" + applicationId %}
-{% set title = "" %}
-{% set formaction = applicationPath + "/work-history-2/answer" %}
-
-{% block pageNavigation %}
-  {% if referrer %}
-    {{ govukBackLink({
-      href: referrer
-    }) }}
-  {% else %}
-    {{ govukBackLink({
-      href: "/application/" + applicationId,
-      text: "Back to application"
-    }) }}
-  {% endif %}
-{% endblock %}
-
-{% block content %}
-  {% set noWorkHistoryHtml %}
-    {{ govukCharacterCount({
-      label: {
-        text: "Tell us why you’ve been out of the workplace"
-      },
-      hint: {
-        text: "For example, ‘I have been unemployed’, ‘I have been volunteering’ or ‘I have been caring for a child’"
-      },
-      threshold: 90,
-      maxwords: 400
-    } | decorateApplicationAttributes(["work-history", "missing"])) }}
-  {% endset -%}
-
-  <div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds">
-  <form{% if formaction %} action="{{ formaction }}"{% endif %} method="post">
-    {{ govukRadios({
-      fieldset: {
-        legend: {
-          classes: "govuk-fieldset__legend--xl",
-          text: "Do you have any work history?",
-          isPageHeading: true
-        }
-      },
-      items: [{
-        value: "yes",
-        text: "Yes"
-      }, {
-        value: "full-time-education",
-        text: "No, I have always been in full time education"
-      }, {
-        value: "no",
-        text: "No, I do not have any work history",
-        conditional: {
-          html: noWorkHistoryHtml
-        }
-      }]
-    } | decorateApplicationAttributes(["work-history-decision"])
-    ) }}
-
-    {{ govukButton({
-      text: "Continue"
-    }) }}
-  </form>
-  <div>
-  </div>
-{% endblock %}
