+++ app/views/application/work-history/index.njk
@@ -1,6 +1,6 @@
-{% extends "_form.njk" %}
+{% extends "_layout.njk" %}
 {% set applicationPath = "/application/" + applicationId %}
-{% set title = "Work history" %}
+{% set title = "" %}
 {% set formaction = applicationPath + "/work-history/answer" %}
 
 {% block pageNavigation %}
@@ -16,30 +16,51 @@
   {% endif %}
 {% endblock %}
 
-{% block primary %}
-  <p class="govuk-body">Training providers need a complete picture of your work history, including time out of the workplace, to safeguard children.</p>
-  <p class="govuk-body">Breaks in work history will not impact your application if you have a reasonable explanation for them (for example, parenting responsibilities, redundancy, illness or personal reasons such as study or travel).</p>
-  {{ govukRadios({
-    name: "decision",
-    fieldset: {
-      legend: {
-        classes: "govuk-fieldset__legend--m",
-        text: "Do you have any work history to add?"
-      }
-    },
-    items: [{
-      value: "yes",
-      text: "Yes, I have a complete work history",
+{% block content %}
+  {% set noWorkHistoryHtml %}
+    {{ govukCharacterCount({
+      label: {
+        text: "Tell us why you’ve been out of the workplace"
+      },
       hint: {
-        text: "You’ll be able to give details of any breaks once you’ve entered your employment history"
-      }
-    }, {
-      value: "no",
-      text: "No, I do not have any work history"
-    }]
-  }) }}
+        text: "For example, ‘I have been unemployed’, ‘I have been volunteering’ or ‘I have been caring for a child’"
+      },
+      threshold: 90,
+      maxwords: 400
+    } | decorateApplicationAttributes(["work-history", "missing"])) }}
+  {% endset -%}
 
-  {{ govukButton({
-    text: "Continue"
-  }) }}
+  <div class="govuk-grid-row">
+  <div class="govuk-grid-column-two-thirds">
+  <form{% if formaction %} action="{{ formaction }}"{% endif %} method="post">
+    {{ govukRadios({
+      fieldset: {
+        legend: {
+          classes: "govuk-fieldset__legend--xl",
+          text: "Do you have any work history?",
+          isPageHeading: true
+        }
+      },
+      items: [{
+        value: "yes",
+        text: "Yes"
+      }, {
+        value: "full-time-education",
+        text: "No, I have always been in full time education"
+      }, {
+        value: "no",
+        text: "No, I do not have any work history",
+        conditional: {
+          html: noWorkHistoryHtml
+        }
+      }]
+    } | decorateApplicationAttributes(["work-history-decision"])
+    ) }}
+
+    {{ govukButton({
+      text: "Continue"
+    }) }}
+  </form>
+  <div>
+  </div>
 {% endblock %}
