--- app/views/application/work-history/job.njk
@@ -1,175 +1,0 @@
-{% extends "_form.njk" %}
-
-{% set applicationPath = "/application/" + applicationId %}
-{% set parent = "Work history" %}
-{% set action = "edit" if applicationValue(["work-history", id]) else "add" %}
-{% set title = (action | capitalize) + " job" %}
-
-{% block pageNavigation %}
-  {% if referrer %}
-    {{ govukBackLink({
-      href: referrer
-    }) }}
-  {% else %}
-    {{ govukBackLink({
-      href: applicationPath + "/work-history"
-    }) }}
-  {% endif %}
-{% endblock %}
-
-{% block primary %}
-  <p class="govuk-body">Include all your paid work in this section.</p>
-  <p class="govuk-body">Do not include voluntary or unpaid experience – we’ll ask you about that later.</p>
-  {{ govukInput({
-    label: {
-      text: "Job title",
-      classes: "govuk-label--m"
-    }
-  } | decorateApplicationAttributes(["work-history", id, "role"])) }}
-
-  {{ govukInput({
-    label: {
-      text: "Name of employer",
-      classes: "govuk-label--m"
-    }
-  } | decorateApplicationAttributes(["work-history", id, "org"])) }}
-
-  {% set partTimeHtml %}
-    {{ govukCharacterCount({
-      label: {
-        text: "Give details about your working pattern",
-        classes: "govuk-label--s"
-      },
-      hint: {
-        text: "For example: ‘20 hours per week’"
-      },
-      maxwords: 50,
-      threshold: 75,
-      rows: 2
-    } | decorateApplicationAttributes(["work-history", id, "part-time-details"])) }}
-  {% endset -%}
-
-  {{ govukRadios({
-    fieldset: {
-      legend: {
-        text: "Was this job full time or part time?",
-        classes: "govuk-fieldset__legend--m"
-      }
-    },
-    items: [{
-      text: "Full time"
-    }, {
-      text: "Part time",
-      conditional: {
-        html: partTimeHtml
-      }
-    }]
-  } | decorateApplicationAttributes(["work-history", id, "type"])) }}
-
-  {{ govukDateInput({
-    fieldset: {
-      legend: {
-        text: "Start date",
-        classes: "govuk-fieldset__legend--m"
-      }
-    },
-    hint: {
-      text: "For example, 5 2018"
-    },
-    namePrefix: id + "-start-date",
-    items: [{
-      label: "Month",
-      name: "month",
-      value: (applicationValue(["work-history", id, "start-date"]) or start or "") | date('L'),
-      classes: "govuk-input--width-2"
-    }, {
-      label: "Year",
-      name: "year",
-      value: (applicationValue(["work-history", id, "start-date"]) or start or "") | date('yyyy'),
-      classes: "govuk-input--width-4"
-    }]
-  } | decorateApplicationAttributes(["work-history", id, "start-date"])) }}
-
-  {{ govukDateInput({
-    fieldset: {
-      legend: {
-        text: "End date (leave blank if this is your current role)",
-        classes: "govuk-fieldset__legend--m"
-      }
-    },
-    hint: {
-      text: "For example, 5 2019"
-    },
-    namePrefix: id + "-end-date",
-    items: [{
-      label: "Month",
-      name: "month",
-      value: (applicationValue(["work-history", id, "end-date"]) or end or "") | date('L'),
-      classes: "govuk-input--width-2"
-    }, {
-      label: "Year",
-      name: "year",
-      value: (applicationValue(["work-history", id, "end-date"]) or end or "") | date('yyyy'),
-      classes: "govuk-input--width-4"
-    }]
-  } | decorateApplicationAttributes(["work-history", id, "end-date"])) }}
-
-  {{ govukCharacterCount({
-    label: {
-      text: "Skills and experience relevant to teaching you gained in this role",
-      classes: "govuk-label--m"
-    },
-    hint: {
-      text: "Give a brief overview of your role and explain how you developed transferable skills like communication, creativity and organisation"
-    },
-    maxwords: 150
-  } | decorateApplicationAttributes(["work-history", id, "description"])) }}
-
-  {{ govukRadios({
-    fieldset: {
-      legend: {
-        text: "Did this job involve working in a school or with children?",
-        classes: "govuk-fieldset__legend--m"
-      }
-    },
-    items: [{
-      text: "Yes"
-    }, {
-      text: "No"
-    }]
-  } | decorateApplicationAttributes(["work-history", id, "worked-with-children"])) }}
-
-  {{ govukInput({
-    type: "hidden",
-    name: "applications[" + applicationId + "][work-history][" + id + "][category]",
-    value: "job"
-  }) }}
-
-  {# Only ask if want to add another if:
-    * not editing an existing item
-    * not referred from a review page (which has an ‘Add another…’ button)
-  #}
-  {% if action != "edit" and not referrer %}
-    <hr class="govuk-section-break govuk-section-break--l govuk-section-break--visible">
-    {{ govukRadios({
-      name: "next",
-      fieldset: {
-        legend: {
-          text: "Do you want to add another job?",
-          classes: "govuk-fieldset__legend--m"
-        }
-      },
-      items: [{
-        value: "add",
-        text: "Yes, I want to add another job"
-      }, {
-        value: "review",
-        text: "No, not at the moment"
-      }]
-    }) }}
-  {% endif %}
-
-  {{ govukButton({
-    text: "Save and continue"
-  }) }}
-{% endblock %}
