+++ app/views/application/work-history/review.njk
@@ -1,10 +1,12 @@
-{% extends "_form-wide.njk" %}
+{% extends "_content-wide.njk" %}
 
 {% set applicationPath = "/application/" + applicationId %}
 {% set formaction = applicationPath %}
 {% set title = "Work history" %}
-{% set referrer = applicationPath + "/work-history/review" %}
 {% set canAmend = true %}
+{% set workHistory = applicationValue(["work-history"]) or {} %}
+{% set history = workHistory | toArray | selectattr("id") | sort(attribute="start-date") %}
+{% set workHistoryDecision = applicationValue(["work-history-decision"]) %}
 
 {% block pageNavigation %}
   {{ govukBackLink({
@@ -14,16 +16,97 @@
 {% endblock %}
 
 {% block primary %}
-  {% include "_includes/review/work-history.njk" %}
 
-  {{ govukCheckboxes({
-    items: [{
-      value: "true",
-      text: "I have completed this section"
-    }]
-  } | decorateApplicationAttributes(["completed", "work-history"])) }}
+{% if workHistoryDecision == "yes" %}
 
-  {{ govukButton({
-    text: "Continue"
+  <div class="govuk-!-width-two-thirds">
+      <p class="govuk-body">Enter all the jobs you've had since you left school.</p>
+
+      <p class="govuk-body">Explain any gaps in your work history. For example, raising children, unemployment or illness.</p>
+    </div>
+
+    {% set buttonText = "Add another job" if history.length > 0 else "Add a job" %}
+    {% set buttonClasses = "govuk-button--secondary" if history.length > 0 %}
+
+    {{ govukButton({
+      text: buttonText,
+      href: "/application/" + applicationId + "/work-history/" + newId,
+      classes: buttonClasses
+      }) }}
+
+
+    {% if history.length >= 1 %}
+      {% include "_includes/review/work-history.njk" %}
+    {% endif %}
+
+{% else %}
+  {% set workHistoryDetailsHtml %}
+    {% if workHistoryDecision == "full-time-education" %}
+
+      {{ govukSummaryList({
+        classes: "app-summary",
+        rows: [{
+          key: {
+            text: "Do you have any work history?"
+          },
+          value: {
+            text: "No, I have always been in full time education"
+          },
+          actions: {
+            items: [{
+              href: applicationPath + "/work-history",
+              text: "Change",
+              visuallyHiddenText: "explanation"
+            }]
+          }
+        }]
+      }) }}
+
+    {% elseif workHistoryDecision == "no" %}
+
+      {{ govukSummaryList({
+        classes: "app-summary",
+        rows: [{
+          key: {
+            text: "Explanation of why you’ve been out of the workplace"
+          },
+          value: {
+            text: (applicationValue(["work-history", "missing"]) or "Not entered") | nl2br | safe
+          },
+          actions: {
+            items: [{
+              href: applicationPath + "/work-history",
+              text: "Change",
+              visuallyHiddenText: "explanation"
+            }]
+          }
+        }]
+      }) }}
+    {% endif %}
+  {% endset %}
+
+  {{ appSummaryCard({
+    classes: "govuk-!-margin-bottom-6",
+    html: workHistoryDetailsHtml
   }) }}
+
+{% endif %}
+
+{% if history.length >= 1 or workHistoryDecision == "no" or workHistoryDecision == "full-time-education"  %}
+
+  <form{% if formaction %} action="{{ formaction }}"{% endif %} method="post">
+    {{ govukCheckboxes({
+      items: [{
+        value: "true",
+        text: "I have completed this section"
+      }]
+    } | decorateApplicationAttributes(["completed", "work-history"])) }}
+
+    {{ govukButton({
+      text: "Save and continue"
+    }) }}
+  </form>
+
+{% endif %}
+
 {% endblock %}
